syscall: print personality switch messages in tcb's output
Personality switch is a rather normal event, it's unclear why it has
been being always treated as a diagnostics/error information. Having
it under qflag makes even little sense, as it wass impossible to get
this information when the output is set to files.  With this change,
it's possible to get it with something like

    strace -f -o out --quiet=none PROG

at least.  One caveat is its rather odd format, with square brackets,
"PID=", and dot, but it seems that we're stuck with it forever, as it
traces back for decades, albeit with a prefix change in commit
v4.11~509 ("Consistently use error_msg instead of fprintf(stderr)").

* syscall.c (update_personality): Use printleader/tprintf/line_ended
instead of error_msg.
* tests/bexecve.test: Add --quiet=personality to strace invocations.
* tests/strace-ff.test: Likewise.